requirements.txt
absl-py==1.2.0
albumentations==1.2.1
apex==0.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bashutils==0.0.4
beautifulsoup4==4.12.2
blessings==1.7
cachetools==5.2.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
chardet==4.0.0
cycler==0.10.0
DCN==1.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dominate==2.7.0
easydict==1.9
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fightingcv-attention==1.0.0
fonttools==4.28.1
future==0.18.2
google==3.0.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.0
h5py==3.7.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
imgaug==0.4.0
importlib-metadata==4.12.0
ipython @ file:///tmp/build/80754af9/ipython_1648799197050/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.3.1
kornia==0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
networkx==2.6.3
numpy==1.20.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
OpenEXR==1.3.9
packaging==21.3
pandas==1.4.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
path==16.2.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.2
plyfile==0.7.4
proglog==0.1.9
progressbar2==3.55.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2022.1
PyWavelets==1.2.0
PyYAML==5.4.1
qudida==0.0.4
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.1
scikit-learn==1.1.2
scipy==1.7.2
setuptools-scm==6.3.2
Shapely==1.8.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.4.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.8.1
torchsummary==1.5.1
torchvision==0.9.1
tornado==6.2
tqdm==4.59.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.2.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.8.1
